+++ ubo_app/utils/apt.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import asyncio
+import subprocess
 
 from ubo_app.logging import logger
 
@@ -53,18 +54,19 @@ def uninstall_package(package_name: str) -> None:
         raise ValueError(msg)
 
 
-def _is_package_installed(package_name: str) -> bool:
-    """Check if a package is installed using the APT package manager."""
-    import apt  # pyright: ignore [reportMissingModuleSource]
-
-    cache = apt.Cache()
-    if package_name in cache:
-        pkg = cache[package_name]
-        return pkg.is_installed
-
-    return False
-
-
 async def is_package_installed(package_name: str) -> bool:
     """Asynchronously check if a package is installed using the APT package manager."""
-    return await asyncio.to_thread(_is_package_installed, package_name)
+    try:
+        process = await asyncio.create_subprocess_exec(
+            'dpkg-query',
+            '-W',
+            '-f=${Status}',
+            package_name,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
+        )
+        stdout, _ = await process.communicate()
+    except subprocess.SubprocessError:
+        return False
+    else:
+        return b'install ok installed' in stdout